--- server.ts
@@ -1,5 +1,0 @@
-import {createServer} from 'http';
-
-
-const PORT = process.env.PORT;
-server.listen(PORT, () => console.log(`Server is running on ${PORT}`))
